rename utility files/change some integrated tests to unit tests
