+++ README.md
@@ -61,8 +61,27 @@ python3 run.py db upgrade
 7. Ask Reed College CIT for the private configuaration file
 In order to run the app
 You need to create a file named ’config.cfg’ and put it into both the root directory and the tests directory(for testing setup) of the app
-Its content should look like the content in `config_example.cfg`(all the values need to be set up)
+Its content should look like this(all the values need to be set up)
+```
+[login_manager]
+SESSION_PROTECTION = 
+LOGIN_VIEW = 
 
+[logger]
+FILEPATH = 
+LEVEL = 
+
+[key]
+SECRET_KEY = 
+
+[app]
+DEBUG = True
+SQLALCHEMY_DATABASE_URI = 
+
+[url]
+STUDENT_API_URL = 
+CLASS_API_URL = 
+```
 
 8. Run the app `python3 __main__.py`
 
@@ -82,7 +101,7 @@ python3 run.py db migrate
 python3 run.py db upgrade
 ```
 ### 4: Testing
-Tests for the backend code: ` python3 tests_script.py ` in lemur/lemur/tests
+Tests for the backend code: ` nose2 ` in lemur/lemur/tests
 
 Tests for the frontend code: Use any main stream browser to open js_test.html in lemur/lemur/tests
