+++ lemur/tests/test_integration_db_and_models.py
@@ -9,16 +9,16 @@ import unittest
 from lemur import (app, db)
 from lemur import models as m
 import helper_random as r
-from lemur.utility.utility_generate_and_convert import (generate_lab_id,
-                                                        generate_experiment_id,
-                                                        generate_observation_id,
-                                                        generate_class_id)
-from lemur.utility.utility_find_and_get import (get_experiment,
-                                                get_observation,
-                                                get_user,
-                                                get_role,
-                                                get_power,
-                                                get_class)
+from lemur.utility.generate_and_convert import (generate_lab_id,
+                                                generate_experiment_id,
+                                                generate_observation_id,
+                                                generate_class_id)
+from lemur.utility.find_and_get import (get_experiment,
+                                        get_observation,
+                                        get_user,
+                                        get_role,
+                                        get_power,
+                                        get_class)
 ds = db.session
