+++ lemur/tests/test_integration_network.py
@@ -17,22 +17,22 @@ from lemur import app, db
 from lemur import models as m
 from db_scripts.db_populate import populate_db
 import helper_random as r
-from lemur.utility.utility_generate_and_convert import (generate_lab_id,
-                                                        generate_experiment_id,
-                                                        generate_observation_id,
-                                                        generate_class_id,
-                                                        pack_labinfo_sent_from_client)
-from lemur.utility.utility_find_and_get import (lab_exists,
-                                                user_exists,
-                                                class_exists,
-                                                observation_exists,
-                                                get_lab,
-                                                get_experiment,
-                                                get_observation,
-                                                get_user,
-                                                get_role,
-                                                get_power,
-                                                get_class)
+from lemur.utility.generate_and_convert import (generate_lab_id,
+                                                generate_experiment_id,
+                                                generate_observation_id,
+                                                generate_class_id,
+                                                pack_labinfo_sent_from_client)
+from lemur.utility.find_and_get import (lab_exists,
+                                        user_exists,
+                                        class_exists,
+                                        observation_exists,
+                                        get_lab,
+                                        get_experiment,
+                                        get_observation,
+                                        get_user,
+                                        get_role,
+                                        get_power,
+                                        get_class)
 ds = db.session
