+++ requirements.txt
@@ -8,7 +8,6 @@ Flask-Login==0.3.2
 Flask-Migrate==1.8.0
 Flask-Script==2.0.5
 Flask-SQLAlchemy==2.1
-Flask-WTF==0.12
 html5lib==0.9999999
 itsdangerous==0.24
 Jinja2==2.8
@@ -16,6 +15,7 @@ Mako==1.0.4
 MarkupSafe==0.23
 migrate==0.3.8
 mpmath==0.19
+nose2==0.6.5
 pbr==1.10.0
 pep8==1.7.0
 psycopg2==2.6.1
@@ -28,4 +28,3 @@ sqlparse==0.1.19
 Tempita==0.5.2
 visitor==0.1.3
 Werkzeug==0.11.9
-WTForms==2.1
